backend/requirements.txt
groq
python-dotenv
pytesseract
pdf2image
fastapi
uvicorn[standard]
python-multipart
